move mocked components into mocks folder
