+++ RetrospectiveExtension.Frontend/components/__tests__/feedbackCarousel.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { shallow } from 'enzyme';
 import FeedbackCarousel from '../../components/feedbackCarousel';
-import { testColumnProps } from './mocked_components/mockedFeedbackColumn';
+import { testColumnProps } from '../__mocks__/mocked_components/mockedFeedbackColumn';
 
 const mockedProps = {
   feedbackColumnPropsList: [testColumnProps],
